Split the `DeferredTests` file in 3 parts
